+++ functions/percona-5.5.sh
@@ -56,7 +56,7 @@ chkconfig mysqld on'
 OS_Debian_Ubuntu='update-rc.d mysqld defaults'
 OS_command
 cd ..
-/bin/rm -rf percona-server-5.5.42-37.1
+[ -d "$percona_install_dir" ] && /bin/rm -rf percona-server-5.5.42-37.1
 cd ..
 
 # my.cf
